pyproject.toml
[tool.poetry]
name = "daewy-ocr-helper"
version = "0.1.0"
description = "Lightweighted API for OCR of Common Pattern like BelegID, Bezugszeichen, Insitutionskennzeichen etc."
authors = ["Maike Broska <maike.broska@spectrumk.de>"]
packages = [{ include = "ocrhelper", from = "src" }]


[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.118.0"
uvicorn = {extras = ["standard"], version = "^0.32.0"}
numpy = "^2.1.2"
opencv-python-headless = "^4.10.0.84"
scikit-image = "^0.24.0"
scikit-learn = "^1.5.2"
pillow = "^10.4.0"
joblib = "^1.4.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
httpx = "^0.27.2"
ruff = "^0.6.9"

[tool.ruff]
line-length = 100

[tool.poetry.scripts]
mini-ocr-train = "ocrhelper.train:cli"
mini-ocr-api = "ocrhelper.api:cli"
